requirements.txt
Django==1.11.1
django-filter==1.0.2
django-jsonfield==1.0.1
djangorestframework==3.6.2
drf-tracking==1.2.0
facepy==1.0.8
gunicorn==19.7.1
Markdown==2.6.8
psycopg2==2.7.1
pytz==2017.2
requests==2.14.2
six==1.10.0
whitenoise==3.3.0
